// lib/schemas.ts
export const GetProgramResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  description: zod.string(),
  country: zod.string(),
  duration: zod.string(),
  imageUrl: zod.string().nullish(),
  featured: zod.boolean(),
  createdAt: zod.string(),
});

export const UpdateProgramParams = zod.object({
  programId: zod.coerce.number(),
});

export const UpdateProgramBody = zod.object({
  title: zod.string().optional(),
  description: zod.string().optional(),
  country: zod.string().optional(),
  duration: zod.string().optional(),
  imageUrl: zod.string().optional(),
  featured: zod.boolean().optional(),
});